+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/clusternetworkaddressconfigprovider/SniRoutingClusterNetworkAddressConfigProvider.java
@@ -9,9 +9,6 @@ package io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider;
 import java.util.Set;
 import java.util.regex.Pattern;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -35,7 +32,7 @@ import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider
  * The following configuration is required:
  * <ul>
  *    <li>{@code bootstrapAddress} a {@link HostPort} defining the host and port of the bootstrap address.</li>
- *    <li>Either {@code brokerAddressPattern} (deprecated) or ${@code advertisedBrokerAddressPattern} must be set (but not both).  These properties contain an address
+ *    <li>${@code advertisedBrokerAddressPattern} must be set. This property contains an address
  *    pattern used to advertise broker addresses.  It is addresses made from this pattern that are returned to the kafka client in the Metadata response so must be
  *    resolvable by the client.  Optionally these properties can specify a port which will be advertised to the clients.  One pattern is supported: {@code $(nodeId)}
  *    which interpolates the node id into the address.
@@ -128,10 +125,6 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
      */
     public static class SniRoutingClusterNetworkAddressConfigProviderConfig {
 
-        private static final Logger LOGGER = LoggerFactory.getLogger(SniRoutingClusterNetworkAddressConfigProviderConfig.class);
-        private static final String ADVERTISED_BROKER_ADDRESS_PATTERN = "advertisedBrokerAddressPattern";
-        public static final String BROKER_ADDRESS_PATTERN = "brokerAddressPattern";
-
         private final HostPort bootstrapAddress;
 
         @JsonIgnore
@@ -140,50 +133,35 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
         private final Pattern brokerAddressNodeIdCapturingRegex;
         @JsonIgnore
         private final Integer advertisedPort;
-
-        // present for serialize/deserialize fidelity
-        @SuppressWarnings("unused")
-        private final String brokerAddressPattern;
         // present for serialize/deserialize fidelity
         @SuppressWarnings("unused")
         private final String advertisedBrokerAddressPattern;
 
         public SniRoutingClusterNetworkAddressConfigProviderConfig(@JsonProperty(required = true) HostPort bootstrapAddress,
-                                                                   @Deprecated(forRemoval = true, since = "0.10.0") @JsonProperty String brokerAddressPattern,
-                                                                   @JsonProperty String advertisedBrokerAddressPattern) {
-            this.brokerAddressPattern = brokerAddressPattern;
+                                                                   @JsonProperty(required = true) String advertisedBrokerAddressPattern) {
             this.advertisedBrokerAddressPattern = advertisedBrokerAddressPattern;
             if (bootstrapAddress == null) {
                 throw new IllegalArgumentException("bootstrapAddress cannot be null");
             }
-            validateOneSpecified(brokerAddressPattern, advertisedBrokerAddressPattern);
-            String brokerAddressPatternProp;
-            String brokerAddressPatternToParse;
-            if (advertisedBrokerAddressPattern != null) {
-                brokerAddressPatternProp = ADVERTISED_BROKER_ADDRESS_PATTERN;
-                brokerAddressPatternToParse = advertisedBrokerAddressPattern;
-            }
-            else {
-                LOGGER.warn("{} is deprecated, replaced by {}", BROKER_ADDRESS_PATTERN, ADVERTISED_BROKER_ADDRESS_PATTERN);
-                brokerAddressPatternProp = BROKER_ADDRESS_PATTERN;
-                brokerAddressPatternToParse = brokerAddressPattern;
+            if (advertisedBrokerAddressPattern == null) {
+                throw new IllegalArgumentException("advertisedBrokerAddressPattern cannot be null");
             }
 
-            PatternAndPort patternAndPort = BrokerAddressPatternUtils.parse(brokerAddressPatternToParse);
+            PatternAndPort patternAndPort = BrokerAddressPatternUtils.parse(advertisedBrokerAddressPattern);
             String brokerAddressPatternPart = patternAndPort.addressPattern();
             advertisedPort = patternAndPort.port().orElse(bootstrapAddress.port());
 
             validatePortSpecifier(brokerAddressPatternPart, s -> {
                 throw new IllegalArgumentException(
-                        brokerAddressPatternProp + " address pattern cannot have port specifier.  Found port : " + s + " within " + brokerAddressPatternPart);
+                        "advertisedBrokerAddressPattern address pattern cannot have port specifier.  Found port : " + s + " within " + brokerAddressPatternPart);
             });
 
             validateStringContainsOnlyExpectedTokens(brokerAddressPatternPart, EXPECTED_TOKEN_SET, tok -> {
-                throw new IllegalArgumentException(brokerAddressPatternProp + " contains an unexpected replacement token '" + tok + "'");
+                throw new IllegalArgumentException("advertisedBrokerAddressPattern contains an unexpected replacement token '" + tok + "'");
             });
 
             validateStringContainsRequiredTokens(brokerAddressPatternPart, EXPECTED_TOKEN_SET, tok -> {
-                throw new IllegalArgumentException(brokerAddressPatternProp + " must contain at least one nodeId replacement pattern '" + tok + "'");
+                throw new IllegalArgumentException("advertisedBrokerAddressPattern must contain at least one nodeId replacement pattern '" + tok + "'");
             });
 
             this.bootstrapAddress = bootstrapAddress;
@@ -191,21 +169,12 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
             this.brokerAddressNodeIdCapturingRegex = BrokerAddressPatternUtils.createNodeIdCapturingRegex(brokerAddressPatternPart);
         }
 
-        private static void validateOneSpecified(String brokerAddressPattern, String advertisedBrokerAddressPattern) {
-            if (brokerAddressPattern != null && advertisedBrokerAddressPattern != null) {
-                throw new IllegalArgumentException("brokerAddressPattern and advertisedBrokerAddressPattern cannot both be specified");
-            }
-            if (brokerAddressPattern == null && advertisedBrokerAddressPattern == null) {
-                throw new IllegalArgumentException("brokerAddressPattern and advertisedBrokerAddressPattern are both null");
-            }
-        }
-
         public HostPort getBootstrapAddress() {
             return bootstrapAddress;
         }
 
-        public String getBrokerAddressPattern() {
-            return brokerAddressPattern;
+        public String getAdvertisedBrokerAddressPattern() {
+            return advertisedBrokerAddressPattern;
         }
 
         public int getAdvertisedPort() {